refactor: Adjust DTO between requests and responses
