+++ src/Application/DTOs/Request/CreateOrderDTO.cs
@@ -1,5 +1,4 @@
 using System.ComponentModel.DataAnnotations;
-using Application.Errors;
 
 namespace Application.DTOs
 {
@@ -14,15 +13,5 @@ namespace Application.DTOs
 
         [Required]
         public required List<OrderItemDTO> OrderItems { get; set; }
-    }
-
-    public class OrderItemDTO
-    {
-        [Required]
-        public long ItemId { get; set; }
-
-        [Required]
-        [Range(1, int.MaxValue, ErrorMessage = ErrorMsg.InvalidQuantity)]
-        public int Quantity { get; set; }
     }
 }
