+++ src/Application/DTOs/Response/OrderItemResponseDTO.cs
@@ -1,0 +1,9 @@
+namespace Application.DTOs
+{
+    public class OrderItemResponseDTO
+    {
+        public long ItemId { get; set; }
+        public int Quantity { get; set; }
+        public decimal PriceCents { get; set; }
+    }
+}
