+++ src/Application/DTOs/Response/OrderResponseDTO.cs
@@ -7,11 +7,4 @@ namespace Application.DTOs
         public decimal TotalPriceCents { get; set; }
         public required List<OrderItemResponseDTO> OrderItems { get; set; }
     }
-
-    public class OrderItemResponseDTO
-    {
-        public long ItemId { get; set; }
-        public int Quantity { get; set; }
-        public decimal PriceCents { get; set; }
-    }
 }
